eval/eval_agent/requirements.txt
pg8000
cloud-sql-python-connector[pg8000]
python-dotenv
google-cloud-secret-manager
requests
GitPython
google-adk
pydantic
